Moved cli parser to new Main class Cleaned Client class Packets are now writen synchronously to avoid packet collision
